% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turku_vaesto_ika_2010_2019}
\alias{turku_vaesto_ika_2010_2019}
\title{Väestörakenne: Väestö iän (1-v.) mukaan kunnan osa-alueittain, 2010-2019}
\format{
A data frame with 133620 observations and 5 variables
\describe{
  \item{Osa-alue}{District number and name, e.g. 853 Turku}
  \item{Vuosi}{Year, 2010-2019}
  \item{Ikä}{Inhabitants' ages in 1 year intervals}
  \item{Tiedot}{Data gathering date, usually 31.12.}
  \item{value}{Number of inhabitants}
  }
}
\source{
\href{https://dev.turku.fi/datasets/pienaluetilastot/}{Turun kaupungin pienaluetilastot}
}
\usage{
data(turku_vaesto_ika_2010_2019)
}
\description{
Population demographics: Population in different city districts
divided by age in years 2010-2019
}
\examples{
data(turku_vaesto_ika_2010_2019)
}
\keyword{datasets}
