% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turku_vaesto_sukupuoli_kieli_2010_2019}
\alias{turku_vaesto_sukupuoli_kieli_2010_2019}
\title{Väestörakenne: Väestö sukupuolen ja kielen mukaan kunnan osa-alueittain, 2010-2019}
\format{
A data frame with 15720 observations and 6 variables
\describe{
  \item{Osa-alue}{District number and name, e.g. 853 Turku}
  \item{Vuosi}{Year, 2010-2019}
  \item{Sukupuoli}{Sex}
  \item{Kieli}{Language}
  \item{Tiedot}{Data gathering date, usually 31.12.}
  \item{value}{Number of inhabitants}
  }
}
\source{
\href{https://dev.turku.fi/datasets/pienaluetilastot/}{Turun kaupungin pienaluetilastot}
}
\usage{
data(turku_vaesto_sukupuoli_kieli_2010_2019)
}
\description{
Population demographics: Population in different city districts
divided by sex and language in years 2010-2019
}
\examples{
data(turku_vaesto_sukupuoli_kieli_2010_2019)
}
\keyword{datasets}
