% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize_data}
\alias{visualize_data}
\title{Visualize data}
\usage{
visualize_data(city = "turku", level = 3, dataset = NULL)
}
\arguments{
\item{city}{The desired city. Options are Helsinki, Espoo, 
Vantaa, Turku, Tampere, Oulu}

\item{level}{The desired administrative level. Options are 1, 2 and 3,
where 1 is the lowest level (= largest areas) and 3 is the highest level 
(= smallest areas)}

\item{dataset}{a dataset containing suitable data}
}
\value{
sf object
}
\description{
Visualize local level data on a choropleth map.
}
\examples{
\dontrun{
visualize_data(city = "turku", level = 3, dataset = turku_vaesto_ika_2010_2019)
}

}
