% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_areas.R
\name{get_areas}
\alias{get_areas}
\title{Get city areas}
\usage{
get_areas(city = "turku", level = 1)
}
\arguments{
\item{city}{The desired city. Options are Helsinki, Espoo, 
Vantaa, Turku, Tampere, Oulu}

\item{level}{The desired administrative level. Options are 1, 2 and 3,
where 1 is the lowest level (= largest areas) and 3 is the highest level 
(= smallest areas)}
}
\value{
sf object
}
\description{
Get city areas
}
